Move resume tokens into process lifecycle
Session resume tokens control synchronous shutdown and orphan reaping, so they belong with process lifecycle code instead of generic utilities.

Constraint: Keep cleanup PRs narrow and preserve existing shutdown behavior.

Rejected: Leave resume-token code in src/utils | Its only production callers are process and Chrome lifecycle paths.

Confidence: high

Scope-risk: narrow

Directive: Process shutdown persistence helpers should live under src/core/process.

Tested: npm test -- --runTestsByPath tests/core/process/session-resume-token.test.ts tests/core/process/sync-shutdown.test.ts tests/chrome/pid-manager.test.ts --runInBand; npm run build; npm run lint; npm run lint:tier; npm run lint:repo-structure; git diff --check

Not-tested: Full manual workflow_dispatch test matrix.
