+++ src/chrome/pid-manager.ts
@@ -2,7 +2,7 @@ import * as fs from "fs";
 import * as os from "os";
 import * as path from "path";
 import { listMarkers, deleteMarkerFile, verifyChromePidIdentity, OwnershipMarker } from "./ownership-marker";
-import { listAllTokens } from "../utils/session-resume-token";
+import { listAllTokens } from "../core/process/session-resume-token";
 import { getLifecycleBus } from "../core/lifecycle";
 import { cleanupAllStaleSessionsSync as cleanupNetworkBodiesSync } from "../core/network-capture/body-store";
