+++ src/core/process/sync-shutdown.ts
@@ -19,7 +19,7 @@
 
 import { killProcessTree } from '../../chrome/pid-manager';
 import { removeMarker } from '../../chrome/ownership-marker';
-import { shouldKillChromeOnExit } from '../../utils/session-resume-token';
+import { shouldKillChromeOnExit } from './session-resume-token';
 
 const LOG_PREFIX = '[openchrome:sync-shutdown]';
